requirements.txt
# Erevna - Multi-Agent Research Analysis System

langchain>=0.2.0
langchain-google-genai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
fastapi>=0.110.0
uvicorn>=0.29.0
